Refactor submodule in `ops`: `webutils` → `web`
